odstraneny typ size_t

+++ file_io.c
@@ -40,7 +40,7 @@ E_ERROR_TYPE check_file_header( char **source_file )
  * @return uspesnost
  */
  
-E_ERROR_TYPE mmap_file(const char *filename, char **file_pointer, size_t* file_size)
+E_ERROR_TYPE mmap_file(const char *filename, char **file_pointer, unsigned *file_size)
 {
     FILE *f;
     if ( ( f = fopen(filename, "r" ) ) == NULL )
@@ -76,7 +76,6 @@ E_ERROR_TYPE mmap_file(const char *filename, char **file_pointer, size_t* file_s
     }
     fclose( f );
     source_file[num_of_chars] = '\0';   //EOF 
-	
 	*file_pointer = source_file;
 	
 	return E_OK;
