+++ file_io.h
@@ -22,4 +22,4 @@
 
 
 E_ERROR_TYPE check_file_header( char **source_file );
-E_ERROR_TYPE mmap_file(const char* filename, char **file_pointer, size_t* file_size);
+E_ERROR_TYPE mmap_file(const char* filename, char **file_pointer, unsigned* file_size);
